minor improvements to Unicode support
Formerly it was still possible for btcrecover.py to crash if it was fed
certain Unicode inputs which it then had to print out in a warning or
error message, e.g. Unicode filenames. This commit replaces the print()
builtin to protect against this. It also prints out any found Unicode
passwords in xml-entity form instead of repr form, which makes it easy
to view as native text in a browser (when console support is lacking).